requirements/requirements.txt
djangorestframework==3.15.1
Django==5.0.4
django-extensions==3.2.3

# for run server
daphne==4.0.0

# for auth
djangorestframework-simplejwt==5.3.1